Session-1/ADK/pyproject.toml
[tool.poetry]
name = "adk"
version = "0.1.0"
description = ""
authors = ["jggomez <juan.gomez01@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
google-adk = "^1.19.0"
google-genai = "^1.52.0"
python-dotenv = "^1.2.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
